backend/chatbot/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
python-socketio==5.12.1
httpx==0.28.1
redis==5.3.0
prometheus-client==0.22.1
